xignzi006 发布系统代码

+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/ibiz/service/IIbzCaseService.java
@@ -11,7 +11,6 @@ import java.math.BigInteger;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.scheduling.annotation.Async;
 import com.alibaba.fastjson.JSONObject;
 import org.springframework.cache.annotation.CacheEvict;
 
@@ -30,18 +29,18 @@ public interface IIbzCaseService extends IService<IbzCase>{
     void createBatch(List<IbzCase> list) ;
     boolean update(IbzCase et) ;
     void updateBatch(List<IbzCase> list) ;
-    boolean remove(Long key) ;
-    void removeBatch(Collection<Long> idList) ;
-    IbzCase get(Long key) ;
+    boolean remove(BigInteger key) ;
+    void removeBatch(Collection<BigInteger> idList) ;
+    IbzCase get(BigInteger key) ;
     IbzCase getDraft(IbzCase et) ;
     boolean checkKey(IbzCase et) ;
     boolean save(IbzCase et) ;
     void saveBatch(List<IbzCase> list) ;
     Page<IbzCase> searchDefault(IbzCaseSearchContext context) ;
-    List<IbzCase> selectByModule(Long id) ;
-    void removeByModule(Long id) ;
-    List<IbzCase> selectByLib(Long id) ;
-    void removeByLib(Long id) ;
+    List<IbzCase> selectByModule(BigInteger id) ;
+    void removeByModule(BigInteger id) ;
+    List<IbzCase> selectByLib(BigInteger id) ;
+    void removeByLib(BigInteger id) ;
     /**
      *自定义查询SQL
      * @param sql  select * from table where id =#{et.param}
