+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/ibiz/service/impl/IbzProjectMemberServiceImpl.java
@@ -79,19 +79,19 @@ public class IbzProjectMemberServiceImpl extends ServiceImpl<IbzProjectMemberMap
 
     @Override
     @Transactional
-    public boolean remove(Long key) {
+    public boolean remove(BigInteger key) {
         boolean result=removeById(key);
         return result ;
     }
 
     @Override
-    public void removeBatch(Collection<Long> idList) {
+    public void removeBatch(Collection<BigInteger> idList) {
         removeByIds(idList);
     }
 
     @Override
     @Transactional
-    public IbzProjectMember get(Long key) {
+    public IbzProjectMember get(BigInteger key) {
         IbzProjectMember et = getById(key);
         if(et==null){
             et=new IbzProjectMember();
