+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/ibizpro/service/impl/IBZProStoryModuleServiceImpl.java
@@ -97,19 +97,19 @@ public class IBZProStoryModuleServiceImpl extends ServiceImpl<IBZProStoryModuleM
 
     @Override
     @Transactional
-    public boolean remove(Long key) {
+    public boolean remove(BigInteger key) {
         boolean result=removeById(key);
         return result ;
     }
 
     @Override
-    public void removeBatch(Collection<Long> idList) {
+    public void removeBatch(Collection<BigInteger> idList) {
         removeByIds(idList);
     }
 
     @Override
     @Transactional
-    public IBZProStoryModule get(Long key) {
+    public IBZProStoryModule get(BigInteger key) {
         IBZProStoryModule et = getById(key);
         if(et==null){
             et=new IBZProStoryModule();
@@ -170,22 +170,22 @@ public class IBZProStoryModuleServiceImpl extends ServiceImpl<IBZProStoryModuleM
 
 
 	@Override
-    public List<IBZProStoryModule> selectByRoot(Long id) {
+    public List<IBZProStoryModule> selectByRoot(BigInteger id) {
         return baseMapper.selectByRoot(id);
     }
 
     @Override
-    public void removeByRoot(Long id) {
+    public void removeByRoot(BigInteger id) {
         this.remove(new QueryWrapper<IBZProStoryModule>().eq("root",id));
     }
 
 	@Override
-    public List<IBZProStoryModule> selectByParent(Long id) {
+    public List<IBZProStoryModule> selectByParent(BigInteger id) {
         return baseMapper.selectByParent(id);
     }
 
     @Override
-    public void removeByParent(Long id) {
+    public void removeByParent(BigInteger id) {
         this.remove(new QueryWrapper<IBZProStoryModule>().eq("parent",id));
     }
