+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/service/ITeamService.java
@@ -11,7 +11,6 @@ import java.math.BigInteger;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.scheduling.annotation.Async;
 import com.alibaba.fastjson.JSONObject;
 import org.springframework.cache.annotation.CacheEvict;
 
@@ -30,9 +29,9 @@ public interface ITeamService extends IService<Team>{
     void createBatch(List<Team> list) ;
     boolean update(Team et) ;
     void updateBatch(List<Team> list) ;
-    boolean remove(Long key) ;
-    void removeBatch(Collection<Long> idList) ;
-    Team get(Long key) ;
+    boolean remove(BigInteger key) ;
+    void removeBatch(Collection<BigInteger> idList) ;
+    Team get(BigInteger key) ;
     Team getDraft(Team et) ;
     boolean checkKey(Team et) ;
     Team managePorjectMembers(Team et) ;
